fixed project names
