+++ 001 - Combining Redux and Firebase/package.json
@@ -26,8 +26,8 @@
     "webpack-dev-server": "^1.14.0"
   },
   "dependencies": {
-    "axios": "^0.9.1",
     "babel-preset-stage-1": "^6.1.18",
+    "firebase": "^2.4.2",
     "lodash": "^3.10.1",
     "react": "^0.14.3",
     "react-dom": "^0.14.3",
