+++ 001 - Combining Redux and Firebase/src/actions/index.js
@@ -1,11 +1,28 @@
-import axios from 'axios';
+import Firebase from 'firebase';
+import _ from 'lodash';
+import {
+  FETCH_POSTS,
+  DELETE_POST,
+  CREATE_POST
+} from './types';
 
-export function fetchUsers() {
-  const request = axios.get('http://jsonplaceholder.typicode.com/users');
+const Posts = new Firebase('https://fbredux.firebaseio.com/');
 
-  return (dispatch) => {
-    request.then(({data}) => {
-      dispatch({ type: 'FETCH_PROFILES', payload: data })
+export function fetchPosts() {
+  return dispatch => {
+    Posts.on('value', snapshot => {
+      dispatch({
+        type: FETCH_POSTS,
+        payload: snapshot.val()
+      });
     });
   };
+}
+
+export function createPost(post) {
+  return dispatch => Posts.push(post);
+}
+
+export function deletePost(key) {
+  return dispatch => Posts.child(key).remove();
 }
