+++ 001 - Combining Redux and Firebase/src/actions/types.js
