+++ 001 - Combining Redux and Firebase/src/components/post_item.js
