+++ 001 - Combining Redux and Firebase/src/reducers/posts_reducer.js
