+++ 001 - Combining Redux and Firebase/style/style.css
@@ -1,0 +1,9 @@
+li.list-group-item {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+form {
+  margin: 10px 0px;
+}
