+++ 002 - Basics of Redux Thunk/src/reducers/index.js
@@ -1,8 +1,7 @@
 import { combineReducers } from 'redux';
-import postsReducer from './posts_reducer';
 
 const rootReducer = combineReducers({
-  posts: postsReducer
+  users: (state = [], action) => action.payload || state
 });
 
 export default rootReducer;
