--- 002 - Basics of Redux Thunk/style/style.css
@@ -1,9 +1,0 @@
-li.list-group-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-form {
-  margin: 10px 0px;
-}
